Adicion de los controladores, modificacion de la entidad Habitacion y reservacion, ademas de modificacion de los repositorios Habitacion y reservacion

+++ src/main/java/com/dog_house/repository/HabitacionRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HabitacionRepositorio extends JpaRepository<Habitacion,Integer>{
+public interface HabitacionRepositorio extends JpaRepository<Habitacion,Long>{
     
 }
